requirements.txt
mysql-replication
CMRESHandler==1.0.0
SQLAlchemy==1.4.22
mysqlclient==2.0.3